+++ examples/entry.rs
@@ -27,6 +27,7 @@ impl FactoryComponent for Counter {
     type CommandOutput = ();
     type ParentInput = AppMsg;
     type ParentWidget = gtk::Box;
+    type Index = DynamicIndex;
 
     view! {
         gtk::Button {