+++ examples/factory.rs
@@ -29,6 +29,7 @@ impl FactoryComponent for Counter {
     type CommandOutput = ();
     type ParentInput = AppMsg;
     type ParentWidget = gtk::Box;
+    type Index = DynamicIndex;
 
     view! {
         root = gtk::Box {